Simplified
Added: @includes breakpoints for responsive builds

Deleted: much of the parts section. I found that I liked to start with one simple source file and then break it out later to separate file based on the project. No need to box myself in early with file name I may or may not use.